Chore/fixes3 (#151)
* force babel env according to aden run mode

* do not build bundles for empty routes

* resolve babel presets with fallback to aden node_modules

* babel preset resolve test

* fix test

* test cover loading presets from aden node_modules

* attempt to fix win spawn issue

* fix doc links

* kill children before done

* ensure anakin kills all children

* upgrade anakin, fix log parser stream detach, update chalk module

* specify node versions

* resolve babel plugins to aden node_modules

* resolve babel plugins to bundled aden node_modules

* remove unnecessary .babelrc

* upgrade anakin for darwin

* attribution

* cleanup

* fix log stream detach, use close event on children

* update js docs

+++ test/lib/spawn.js
@@ -1,17 +1,45 @@
-const nativeSpawn = require('child_process').spawn;
+const os = require('os');
+const childProcess = require('child_process');
+const nativeSpawn = childProcess.spawn;
+const exec = childProcess.exec;
 
-const children = [];
+let children = [];
 const spawn = (...rest) => {
   const child = nativeSpawn(...rest);
   children.push(child);
   return child;
 };
 
-const anakin = () => {
-  children.forEach((child) => {
-    child.kill('SIGINT');
-  });
+const anakin = (done) => {
+  const registeredChildren = children;
+  children = [];
+
+  return Promise.all(registeredChildren
+    .map((child) => new Promise((resolve) => {
+      if (child.killed || [0, 1].includes(child.exitCode)) {
+        resolve();
+        return;
+      }
+      child.on('close', () => setTimeout(() => resolve(), 500))
+      if (os.platform() === 'win32') {
+        // https://stackoverflow.com/a/28163919
+        // Works best so far.
+        exec(`taskkill /pid ${child.pid} /T /F`);
+      } else {
+        child.kill('SIGINT');
+        setTimeout(() => child.kill('SIGKILL'), 1000);
+      }
+    })))
+    .then(() => {
+      if(typeof done === 'function') {
+        done();
+      }
+    });
 };
+
+process.on('error', () => {
+  anakin();
+});
 
 process.on('exit', () => {
   anakin();
